Cookie login things.

+++ includes/db.php
@@ -1,5 +1,8 @@
 <?php
-
+	//Basicly make db.php able to include settings.php from anywhere in the project.
+	require_once dirname(__FILE__).'/settings.php';
+	if ($settings['debug']) {echo "<br><br>".dirname(__FILE__)."<br><br>";}
+	
 	//This class connects to the database and it's functions can be used for comunicating with database. This is done so database code can be easily switched to run some other type like PDO...
 	
 	class DAO {
@@ -9,7 +12,7 @@
 				$this->connection = new mysqli("localhost",$db_username,$db_password,$db_database);
 			} catch (Exception $e) {
 				echo "Database could not connect";
-				var_dump($e);
+				var_dump($e);		
 				die();
 			}
 		}
@@ -37,9 +40,9 @@
 		}
 
 		//Delete link from database
-		public function deleteLink($id) {
-			$stmt = $this->connection->prepare("DELETE FROM links WHERE id=?");
-			$stmt->bind_param("s", $id); // Bind username varaible to sql statement above.
+		public function deleteLink($owner,$id) {
+			$stmt = $this->connection->prepare("DELETE FROM links WHERE id=? AND UserID=?");
+			$stmt->bind_param("ii", $id,$owner); // Bind username varaible to sql statement above.
 			$result = $stmt->execute(); //Run sql 
 			$stmt->close();
 
@@ -47,7 +50,7 @@
 
 		//Check if username and password match and are found from db.
 		public function checkLogin($username,$password) {
-			$password = password_hash($password, PASSWORD_BCRYPT);
+
 			$stmt = $this->connection->prepare("SELECT password,disabled FROM users WHERE username=?");
 			$stmt->bind_param("s", $username); // Bind username varaible to sql statement above.
 			$result = $stmt->execute(); //Run sql 
@@ -55,10 +58,44 @@
 			$stmt->close();
 			
 			//If user's password matches and account is not disabled return true.
-			if ($password == $result[0] && $result[1] != "true") {
+			if (password_verify($password,$result[0]) && $result[1] != "true") {
 				return true;
 			}
 			return false;
+		}
+
+		/*
+		This is to create a secure string that is stored in database and in user's device as cookie.
+		This might not be the best implementation but it works for this school project. When users comes back to the site this string can be used to authentificate the login.
+		*/
+		public function updateSecureString($id) {
+
+			$lenght = 30;
+			$letters = "ABCDeFGHIJKLMNOPQRSTUVXYZO123456789abcdefghijklmnopqrstuvxyzo";
+			$letters = str_split($letters,1);
+			$string = ""; //List for letters of secure string.
+			$max = count($letters)-1; //max lenght for the random.
+
+			//Add x amout of letters to $parts
+			for ($i = 0; $i < $lenght; ++$i) {
+				$string .= $letters[random_int(0, $max)]; //Add random letter to list
+			}
+
+			$stmt = $this->connection->prepare("UPDATE users SET auth=? WHERE id=?");
+			$stmt->bind_param("si",$string,$id); // Bind username varaible to sql statement above.
+			$result = $stmt->execute(); //Run sql
+			$stmt->close();
+
+			return $string;
+		}
+
+		public function checkAuth($auth) {
+			$stmt = $this->connection->prepare("SELECT ID,username FROM users WHERE auth=?");
+			$stmt->bind_param("s", $auth); // Bind username varaible to sql statement above.
+			$result = $stmt->execute(); //Run sql 
+			$result = $stmt->get_result()->fetch_row(); //Get results but only the first row. (No need for more because there can be only one user with same name.)
+			$stmt->close();
+			return $result[0];
 		}
 
 		//Check user's ID
@@ -95,19 +132,21 @@
 				$stmt->execute(); //Run sql
 				$stmt->close();
 
-			} else {
-				echo "Username is taken";
-				return false;
+				return true;
 			}
-			echo $found;
+
+			return false;
+
 		}
 
 	}
 
 
-$dao_obj = new DAO("launchpage","","");
-$dao_obj->createAccount("Testi","1234");
-var_dump($dao_obj->checkLogin("Testi","1234"));
+$dao_obj = new DAO($db_name,$db_user,$db_password);
+//$dao_obj->createAccount("Testi","1234");
+//var_dump($dao_obj->checkLogin("cmdtvt","cookie"));
+//echo password_hash("cookie",PASSWORD_DEFAULT);
+
 //var_dump($dao_obj->checkLogin("system","system"));
